Make exporting (and in general, saving) honor the selected filter.
This auto-appends the extension of the selected filter to the filename,
and avoids displaying the extension to the user in the chooser dialog.
This actually unifies the Win32 and UNIX behavior quite a bit and
should reduce some confusion as well.  As a result of this, we now
just have one export option in the menu.