Fix (again) Win Vista/7 SaveAs dialog box file type filters
I thought I had fixed this in changeset:975, but apparently not. It
seems that WinXP wants the "CustomFilter" to have the default,
and Filter[] to have a list of other filter types, not including the
default. Win Vista/7 seem to want the default in the Filter[] list
as well. This makes the code do the right thing on the right platform.