Add memory detail editor
This adds an "Edit" option to the right-click context menu for a memory,
which opens a dialog window allowing somewhat easier atomic editing of a
memory's values. I think this will provide an easier way to expose some
more complicated and radio-specific features of the memory objects,
such as split tones and perhaps even D-STAR attributes.

Right now, it exposes any Memory.extra features after the default
properties of a memory, which will cause the window to grow unbounded.
This may be resolved with a ScrolledWindow, or my favorite: tabs.
Also note that the Memory.extra object is expected to be a flat single-level
RadioSettingGroup right now.

Mostly addresses #165