[developer] Control fontsize of hex diff/dump display - #1681
Add user specification of the font size to use for the hex diff/dump
display, without affecting any other Chirp displays.
This is done with a new chirp.config integer property diff_fontsize
in a new [developer] section. The default is unchanged at 11.
Values between 4 and 144 are accepted; others result in the default.

Add a new file README.developers at the top level of the repository
to document chirp.config [developer] section properties.

#1681