Fix chirp.logger module (#2347)
This patch addresses some issues with the chirp.logger module in order
to use it in chirpc and chirpw.  It prints the version string when
logging at DEBUG level.

In addition, it embraces and extends the "debug.log" support that was
present in chirpw, ensuring all log messages and output end up in that
file when running on Win32 or MacOS hosts.  It also allows that feature
to be tested on Linux by setting CHIRP_DEBUG_LOG in the environment.