[IC-7000] Add Split Duplex fixes #3101
Independent receive and transmit frequencies can be set for each IC-7000 memory channel.  This is enabled when Duplex is set to split.  The transmit frequency is entered in the Offset field.

This patch exposes the Offset field for all memories.  Offset changes to memories with plus, minus, or no split are silently ignored; the IC-7000 does not support per channel offset.