[KT8900D] Add Alias for Surecom S-KT8900D
Owners of the Surecom S-KT8900D mobile radio can easily be confused as to which
vendor/model to choose in order to program their radio because it is a variant
of the QYT KT7900D mobile radio and not the QYT KT8900D as you would assume.

This patch adds an alias for the Surecom KT-8900D which will automatcally
choose QYT KT7900D as the vendor/model.

related to issue #4697